quant-research-pipeline/repo/pyproject.toml
[project]
name = "strategy-fx_cookbook"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pandas==2.3.3",
  "numpy==2.4.2",
  "scipy==1.17.0",
  "pyyaml==6.0.3",
  "pydantic==2.12.5",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
  "pytest-cov==5.0.0",
  "hypothesis==6.151.9",
  "ruff==0.15.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
